added: capability to specify Dockle version
Updated github action to include a new optional argument called
`dockle-version`, which allows you to specify the version of Dockle
that you want to use for the scan. If the value is not provided, it will
default to `latest`.

Added new test, and updated the `lib` folder to include the changes.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>